Rename everything match JSON to Json (in namespaces and class names)
