+++ README.md
@@ -5,7 +5,7 @@
 JSON is a class that provides tools to manage your data in JSON format. You can use it as array or object handler, and make type conversions between these three types (JSON, array, object). The indexes, for example, can be accessed using nested keys, separated by dots (e.g. apps.browsers.chrome).
 
 ## Documentation
-See [the wiki](https://github.com/MAChitgarha/JSON/wiki).
+See [the wiki](https://github.com/MAChitgarha/Json/wiki).
 
 ## Installing
 You can easily install it with Composer:
