+++ tests/unit/InvalidJsonExceptionTest.php
@@ -1,22 +1,22 @@
 <?php
 
 /**
- * Unit tests for MAChitgarha\Component\JSON class.
+ * Unit tests for MAChitgarha\Component\Json class.
  *
  * Go to the project's root and run the tests in this way:
  * phpunit --bootstrap vendor/autoload.php tests/unit
  *
- * @see MAChitgarha\Component\JSON
+ * @see MAChitgarha\Component\Json
  */
-namespace MAChitgarha\UnitTest\JSON;
+namespace MAChitgarha\UnitTest\Json;
 
 use PHPUnit\Framework\TestCase;
-use MAChitgarha\Component\JSON;
-use MAChitgarha\Exception\JSON\InvalidJsonException;
+use MAChitgarha\Component\Json;
+use MAChitgarha\Json\Exception\InvalidJsonException;
 
 class InvalidJsonExceptionTest extends TestCase
 {
-    /** @var JSON */
+    /** @var Json */
     protected $json;
 
     protected function setUp()
